Refactor to use grid display in listing cards, consolidated to one listing card, minor changes to card display, conditional on window size

+++ client/src/pages/account.tsx
@@ -1,8 +1,7 @@
 import {useState, useEffect} from "react";
 import {NewListing} from '../types/types'
-import OwnListingsCard from '../components/accounts/OwnListingsCard'
-import FavListingsCard from "../components/accounts/FavListingsCard";
-import NewListingModal from "../components/accounts/NewListingModal";
+import ListingCard from '../components/accounts/ListingCard'
+import ListingModal from "../components/accounts/ListingModal";
 import axios from '../utils/axios';
 import swal from 'sweetalert';
 
@@ -86,61 +85,56 @@ const Account = () => {
         setSelectedListing(null);
     };
 
-    const handleUnfavorite = (listingId: number) => {
+    const updateFavorite = (listing: NewListing, listingId: number, favorite: boolean) => {
         const prevFavListings = favListings;
         const prevFavListingsIds = favListingsIds;
-
-        setFavListings(prevFavListings.filter((listing) => listing.id !== listingId));
-        setFavListingsIds(prevFavListingsIds.filter((id) => id !== listingId));
-
-        axios.delete('/users/favListings', {withCredentials: true, data: {favListings: [listingId]}}).catch((error) => {
-            setFavListings(prevFavListings);
-            setFavListingsIds(prevFavListingsIds);
-            console.error('Error unfavoriting listing:', error);
-            swal({
-                text: "Unable to unfavorite listing",
-                icon: "warning",
-            })
-        });
-    };
-
-    const handleFavorite = (listing: NewListing ,listingId: number) => {
-        const prevFavListings = favListings;
-        const prevFavListingsIds = favListingsIds;
-
-
-        setFavListings([...prevFavListings, listing]);
-        setFavListingsIds([...prevFavListingsIds, listingId]);
-
-        axios.put('/users/favListings', {withCredentials: true, data: {favListings: [listing.id]}}).catch((error) => {
-            setFavListings(prevFavListings);
-            setFavListingsIds(prevFavListingsIds);
-            console.error('Error favoriting listing:', error);
-            swal({
-                text: "Unable to favorite listing",
-                icon: "warning",
-            })
-        });
+        
+        if(favorite) {
+            setFavListings([...prevFavListings, listing]);
+            setFavListingsIds([...prevFavListingsIds, listingId]);
+    
+            axios.put('/users/favListings', {withCredentials: true, data: {favListings: [listing.id]}}).catch((error) => {
+                setFavListings(prevFavListings);
+                setFavListingsIds(prevFavListingsIds);
+                console.error('Error favoriting listing:', error);
+                swal({
+                    text: "Unable to favorite listing",
+                    icon: "warning",
+                })
+            });
+        } else {
+            setFavListings(prevFavListings.filter((listing) => listing.id !== listingId));
+            setFavListingsIds(prevFavListingsIds.filter((id) => id !== listingId));
+    
+            axios.delete('/users/favListings', {withCredentials: true, data: {favListings: [listingId]}}).catch((error) => {
+                setFavListings(prevFavListings);
+                setFavListingsIds(prevFavListingsIds);
+                console.error('Error unfavoriting listing:', error);
+                swal({
+                    text: "Unable to unfavorite listing",
+                    icon: "warning",
+                })
+            });
+        }
     };
 
     return (
-        <div className="p-8 mt-[4rem]">
+        <div className="p-8 transition-all lg:mx-12 mt-[4rem]">
             {isLoading ? (
                 <div style={{marginTop: '17%', textAlign: 'center'}}>
                     <PulseLoader color="#66CCFF" size={10} /> 
                 </div>
             ) : (
-                <div>
+                <div className="">
                     <p className="text-xl text-gray-700 mb-4">Your listings</p>
                     {ownListings.length > 0 ? (
                         <ul>
                             {ownListings.map((listing) => (
                                 <li key={listing.id} className="mb-2" onClick={() => openModal(listing)}>
-                                    <OwnListingsCard 
+                                    <ListingCard 
                                         listing={listing} 
                                         favListingsIds={favListingsIds} 
-                                        unfavoriteListing={handleUnfavorite} 
-                                        favoriteListing={handleFavorite}
+                                        updateFavorite={updateFavorite}
                                     />
                                 </li>
                             ))}
@@ -153,9 +147,10 @@ const Account = () => {
                         <ul>
                             {favListings.map((listing) => (
                                 <li key={listing.id} className="mb-2" onClick={() => openModal(listing)}>
-                                    <FavListingsCard 
-                                        listing={listing} 
-                                        unfavoriteListing={handleUnfavorite}
+                                    <ListingCard
+                                        listing={listing}
+                                        favListingsIds={favListingsIds}
+                                        updateFavorite={updateFavorite}
                                     />
                                 </li>
                             ))}
@@ -166,13 +161,12 @@ const Account = () => {
                     
                     {/* Modal */}
                     {selectedListing && (
-                        <NewListingModal 
+                        <ListingModal 
                             isOpen={isModalOpen} 
                             onClose={closeModal} 
                             listing={selectedListing}
                             favListingsIds={favListingsIds}
-                            unfavoriteListing={handleUnfavorite} 
-                            favoriteListing={handleFavorite}
+                            updateFavorite={updateFavorite}
                         />
                     )}
                 </div>
